added granularity to methods

+++ internal/configs/app.go
@@ -1,8 +1,8 @@
 package configs
 
 type AppConfig struct {
-	DefaultObservationLimit    int64 `yaml:"defaultObservationLimit"`
-	MaxCorrelationObservations int64 `yaml:"maxCorrelationObservations"`
-	MinCorrelationObservations int64 `yaml:"minCorrelationObservations"`
-	UpdateCorrelationsBuffer   int   `yaml:"updateCorrelationsBuffer"`
+	DefaultObservationLimit    int `yaml:"defaultObservationLimit"`
+	MaxCorrelationObservations int `yaml:"maxCorrelationObservations"`
+	MinCorrelationObservations int `yaml:"minCorrelationObservations"`
+	UpdateCorrelationsBuffer   int `yaml:"updateCorrelationsBuffer"`
 }
